requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
flask-sqlalchemy==3.0.5
flask-jwt_extended==4.5.3
werkzeug==2.3.7
flask-limiter==3.5.0
redis>=4.0.0
stripe==7.8.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
sentry-sdk[flask]==2.35.0
apscheduler==3.10.1
pymongo
pycryptodome
pycryptodomex
cryptography
aiohttp
asyncio
aiofiles
aiohttp-socks
aiohttp-proxy
aiohttp-cors
beautifulsoup4
lxml
html5lib
